+++ devoutils/faker/providers/numbers_provider.py
@@ -8,11 +8,11 @@ from faker.providers import BaseProvider
 class NumbersProvider(BaseProvider):
     """Main class for number provider"""
     @staticmethod
-    def hex(min=0, max=sys.maxsize):
+    def hex(min_value=0, max_value=sys.maxsize):
         """Return hex random number"""
-        return hex(random.randint(min, max))
+        return hex(random.randint(min_value, max_value))
 
     @staticmethod
-    def int(min=0, max=sys.maxsize):
+    def int(min_value=0, max_value=sys.maxsize):
         """Return int random number"""
-        return random.randint(min, max)
+        return random.randint(min_value, max_value)