+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.0.1] - 2019-10-17
-#### Fixed
- * Fixed devo-sdk version required
+## [3.0.0] - 2019-10-17
+#### Changed
+ * Modified base devo-sender from version 3.0.x to 3.2.5
+ * Faker cli add verbose mode to show the events in the console
+ * Add file_name param to define a file to store events in batch mode or 
+ for testing, store in a file but do not send it.
 
 ## [2.0.0] - 2019-07-02
 #### Changed